+++ utils/vectors.c
@@ -6,7 +6,7 @@
 /*   By: bbellavi <bbellavi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/03/13 14:08:59 by bbellavi          #+#    #+#             */
-/*   Updated: 2020/05/14 12:41:43 by bbellavi         ###   ########.fr       */
+/*   Updated: 2020/06/03 00:14:29 by bbellavi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -136,14 +136,14 @@ t_vec	div_vec(t_vec a, int scalar)
 	return (ret);
 }
 
-int		dist_vec(t_vec a, t_vec b)
+double	dist_vec(t_vec a, t_vec b)
 {
 	const int dist = (b.x - a.x) * (b.x - a.x) + (b.y - a.y) * (b.y - a.y);
 	
 	return (sqrt(dist));
 }
 
-int		dist_vec2(t_vec2 a, t_vec2 b)
+double	dist_vec2(t_vec2 a, t_vec2 b)
 {
 	const int dist = (b.x - a.x) * (b.x - a.x) + (b.y - a.y) * (b.y - a.y);
